+++ R/argonColumn.R
@@ -5,6 +5,7 @@
 #' @param ... Any UI element.
 #' @param width Optional. Column width between 1 and 12.
 #' @param center Whether to center column elements or not. FALSE by default.
+#' @param offset Column offset. NULL by default. Importantly, offset + width cannot exceed 12!
 #' 
 #' @examples 
 #' if (interactive()) {
@@ -24,10 +25,10 @@
 #' @author David Granjon, \email{dgranjon@@ymail.com}
 #'
 #' @export
-argonColumn <- function(..., width = NULL, center = FALSE) {
-  
-  colCl <- if(!is.null(width)) paste0("col-sm-", width) else "col-sm"
+argonColumn <- function (..., width = NULL, center = FALSE, offset = NULL) {
+  colCl <- "col-sm"
+  if (!is.null(width)) colCl <- paste0(colCl, " col-sm-", width)
+  if (!is.null(offset)) colCl <- paste0(colCl, " offset-sm-", offset)
   if (center) colCl <- paste0(colCl, " text-center")
-  
   htmltools::tags$div(class = colCl, ...)
 }